+++ backend/dataall/modules/datasets/services/dataset_table.py
@@ -6,7 +6,7 @@ from dataall.db import models, api, permissions, exceptions, paginate
 from dataall.db.api import has_tenant_perm, has_resource_perm, Glossary, ResourcePolicy, Environment
 from dataall.db.models import Dataset
 from dataall.utils import json_utils
-from dataall.modules.datasets.db.table_column_model import DatasetTableColumn
+from dataall.modules.datasets.db.models import DatasetTableColumn
 
 logger = logging.getLogger(__name__)
